streams, pdf, deleting files, finish file upload module

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const { validationResult } = require("express-validator/check");
 const Product = require("../models/product");
+const fileHelper = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
 	res.render("admin/edit-product", {
@@ -24,7 +25,7 @@ exports.postAddProduct = (req, res, next) => {
 			editing: false,
 			hasError: true,
 			product: { title, price, description },
-			errorMessage: 'Attached file is not an image',
+			errorMessage: "Attached file is not an image",
 			validationErrors: [],
 		});
 	}
@@ -127,6 +128,7 @@ exports.postEditProduct = (req, res, next) => {
 			product.price = updatedPrice;
 			product.description = updatedDesc;
 			if (image) {
+				fileHelper.deleteFile(product.imageUrl);
 				product.imageUrl = image.path;
 			}
 			return product.save().then((result) => {
@@ -144,7 +146,16 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
 	const prodId = req.body.productId;
-	Product.deleteOne({ _id: prodId, userId: req.user._id })
+
+	Product.findById(prodId)
+		.then((product) => {
+			if (!product) {
+				return next(new Error("Product not found"));
+			}
+
+			fileHelper.deleteFile(product.imageUrl);
+			return Product.deleteOne({ _id: prodId, userId: req.user._id });
+		})
 		.then(() => {
 			console.log("destroyed");
 			res.redirect("/admin/products");
